Roman Numeral solution using lookup table
This is way less spagetti code and uses a data structure as a lookup
table which would make this easier to maintain over time and migrate to
a kv store or database.  I don't like having to put empty lookups in
however this would be handled in the real world by a library.

Added a test case to handle a data error I found.